+++ mod.ts
@@ -1,6 +1,8 @@
 export type { Config } from "./src/config.ts";
 export { config } from "./src/config.ts";
 
+export { InvalidTimeoutError, MissingApiKeyError } from "./src/errors.ts";
+
 export type {
   AccountApiParameters,
   AccountInformation,